change: move unique to Array.js
