+++ README.md
@@ -4,20 +4,51 @@
 
 ### Table of Contents
 
-- [is](#is)
+- [getFirst](#getfirst)
     -   [Parameters](#parameters)
+- [unique](#unique)
+    -   [Parameters](#parameters-1)
+- [is](#is)
+    -   [Parameters](#parameters-2)
     -   [Examples](#examples)
 - [isEqual](#isequal)
-    -   [Parameters](#parameters-1)
-- [has](#has)
-    -   [Parameters](#parameters-2)
-    -   [Examples](#examples-1)
-- [equal](#equal)
     -   [Parameters](#parameters-3)
-- [unique](#unique)
+- [Object](#object)
+- [has](#has)
     -   [Parameters](#parameters-4)
-- [getByProp](#getbyprop)
+- [equal](#equal)
     -   [Parameters](#parameters-5)
+- [getByProp](#getbyprop)
+    -   [Parameters](#parameters-6)
+
+## getFirst
+
+Get first item in array.
+
+### Parameters
+
+- `array` **[Array](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array)** 
+
+Returns **Any** first item in array
+
+**Meta**
+
+- **since**: 0.0.3
+
+## unique
+
+Get Unique value in array.
+
+### Parameters
+
+- `arrs` **[Array](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array)** 
+- `prop` **[String](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String)** as properties
+
+Returns **[Array](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array)** new array
+
+**Meta**
+
+- **since**: 0.0.2
 
 ## is
 
@@ -48,24 +79,26 @@ Returns **[Boolean](https://developer.mozilla.org/docs/Web/JavaScript/Reference/
 
 Returns **[Boolean](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Boolean)** 
 
+## Object
+
 ## has
 
 checking the property in your object.
 
 ### Parameters
 
-- `obj` **[Object](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Object)** 
+- `obj` **[Object](#object)** 
 - `propertyName` **[String](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String)** 
 
-### Examples
-
-```javascript
-const Person = {name: 'John Lennon', band: 'Beatles', genre: "rock n roll"}
-
-has(Person,'name')
-```
-
 Returns **[Boolean](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Boolean)** 
+
+**Meta**
+
+- **since**: 0.0.2
+
+    const Person = {name: 'John Lennon', band: 'Beatles', genre: "rock n roll"}
+
+    has(Person,'name')
 
 ## equal
 
@@ -73,23 +106,18 @@ Compare 2 Object
 
 ### Parameters
 
-- `objX` **[Object](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Object)** 
-- `objY` **[Object](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Object)** 
+- `objX` **[Object](#object)** 
+- `objY` **[Object](#object)** 
 
-Returns **Bool** This function for simple json object compare
-you can't compare specific JSON like NODE Or
-Dom Object in JSON-Data.
+Returns **Bool** 
 
-## unique
+**Meta**
 
-Get Unique value in array.
+- **since**: 0.0.2
 
-### Parameters
-
-- `arrs` **[Array](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array)** 
-- `prop` **[String](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String)** as properties
-
-Returns **[Array](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array)** new array
+    This function for simple json object compare
+    you can't compare specific JSON like NODE Or
+    Dom Object in JSON-Data.
 
 ## getByProp
 
@@ -97,7 +125,11 @@ Get value of object by property name.
 
 ### Parameters
 
-- `obj` **[Object](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Object)** to find value in this object
+- `obj` **[Object](#object)** to find value in this object
 - `prop` **[String](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String)** property name in obj
 
 Returns **Any** value in object by property name
+
+**Meta**
+
+- **since**: 0.0.2
