+++ src/index.js
@@ -1,3 +1,3 @@
 export { default as Boolean } from "./Boolean";
 export { default as Object } from "./Object";
-export {default as Array} from './Array'
+export { default as Array } from "./Array";
